Separate online IAS functionality into 'ias' feature
