+++ Cargo.toml
@@ -7,15 +7,16 @@ edition = "2018"
 [features]
 default = []
 verbose = []
+ias = ["http", "hyper", "hyper-tls"]
 
 [dependencies]
 anyhow = "1.0"
 base64 = "0.11"
 byteorder = "1.3"
 hex = "0.4"
-http = "0.2"
-hyper = "0.13"
-hyper-tls = "0.4"
+http = { version = "0.2", optional = true }
+hyper = { version = "0.13", optional = true }
+hyper-tls = {version = "0.4", optional = true }
 openssl = "0.10"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
